% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{day8}
\alias{day8}
\title{Compute the start of 8day intervals}
\usage{
day8(x)
}
\arguments{
\item{x}{Date, vector of one or more dates}
}
\value{
a vector of Dates representing the start day
  of the 8-day interval capturing each input date
}
\description{
Compute the start of 8day intervals
}
