% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_woy}
\alias{draw_woy}
\title{Quick and dirty drawing for a prediction map}
\usage{
draw_woy(date = "2016-01-01", species = c("Green", "Hawksbill")[1],
  obs = read_obs(),
  xcast_path = file.path("/mnt/ecocast/projectdata/turtlewatch", "xcast",
  "8D", species))
}
\arguments{
\item{date}{the date to draw (note - please be in 2016)}

\item{species}{character, either Green or Hawksbill}

\item{obs}{sf object, of observations}

\item{preds}{list, a named list of predictor variables (as rasters)}
}
\value{
leaflet map object
}
\description{
Quick and dirty drawing for a prediction map
}
