% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{list_models}
\alias{list_models}
\title{List the model directories}
\usage{
list_models(path = file.path("~", "turtlewatch", "model", "Green"))
}
\arguments{
\item{path}{character the path to models (saved in subdirectories)}
}
\value{
character vector fzero or more fully ualified path descriptions
}
\description{
List the model directories
}
