% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcast.R
\name{list_xcasts}
\alias{list_xcasts}
\title{List the model directories}
\usage{
list_xcasts(path = ".")
}
\arguments{
\item{path}{character the path to models (saved in subdirectories)}
}
\value{
character vector fzero or more fully ualified path descriptions
}
\description{
List the model directories
}
