% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_8day}
\alias{model_8day}
\title{Create maxent models for each 8day week}
\usage{
model_8day(species = c("Green", "Hawksbill")[1], obs = read_obs(),
  preds = read_rasters(), min_obs = 3, n_back = 50, path = ".")
}
\arguments{
\item{species}{character, either Green or Hawksbill}

\item{obs}{sf object, of observations}

\item{preds}{list, a named list of predictor variables (as rasters)}

\item{min_obs}{numeric, the minimum number of observations need to run a model}

\item{n_back}{numeric, the number of background points to select}

\item{path}{character, the output path for the models identified by
first doy of year for the 8-day week.}
}
\value{
a list of zero or more models (some possibly NULL)
  N.B. not every woy will produce a model
}
\description{
Create maxent models for each 8day week
}
