% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obs_bbox}
\alias{obs_bbox}
\title{Retrieve the bounding box of the observations}
\usage{
obs_bbox(x = read_obs(), pad = c(0.05, 0.05))
}
\arguments{
\item{x}{table of observations}

\item{pad}{numeric, 1, 2 or 4 element vector of positive padding to add
to box in degrees.  If a 2 element vector then it should be [x,y],
if a 4 element vector it should be [left, right, bottom, top].
If it is a 1 element vector then it is used in each direction.}
}
\value{
a 4 element numeric vector of [left, right, bottom, top]
}
\description{
Retrieve the bounding box of the observations
}
