% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{read_models}
\alias{read_models}
\title{Read the model directories}
\usage{
read_models(paths = list_models(path))
}
\arguments{
\item{paths}{character the paths to models (saved in subdirectories)}
}
\value{
list of zero or more maxent models
}
\description{
Read the model directories
}
