% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{read_raster}
\alias{read_raster}
\title{Read a raster brick}
\usage{
read_raster(param = c("CHL", "NSST", "SST")[1],
  path = system.file("raster", package = "turtlewatch"))
}
\arguments{
\item{param}{character the name of the parameter - either CHL, SST, NSST}

\item{path}{the path to the rasters}
}
\value{
RasterBrick or NULL
}
\description{
Read a raster brick
}
