% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{read_rasters}
\alias{read_rasters}
\title{Read a series of raster bricks}
\usage{
read_rasters(param = c("CHL", "NSST", "SST"), ...)
}
\arguments{
\item{param}{character vector of the parameters to read.
Default CHL, SST and NSST}

\item{...}{extra arguments for \code{\link{read_raster}}}
}
\value{
named list of bricks
}
\description{
Read a series of raster bricks
}
