% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcast.R
\name{read_xcasts}
\alias{read_xcasts}
\title{Read the model directories}
\usage{
read_xcasts(paths = list_xcasts())
}
\arguments{
\item{paths}{character the paths to models (saved in subdirectories)}
}
\value{
list of zero or more maxent models
}
\description{
Read the model directories
}
