% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcast.R
\name{xcast_8day}
\alias{xcast_8day}
\title{Generate xcast rasters for a given species and year}
\usage{
xcast_8day(species = c("Green", "Hawksbill")[1], year = "2016",
  preds = read_rasters(), model_path = ".", xcast_path = ".")
}
\arguments{
\item{species}{character, either Green or Hawksbill}

\item{preds}{list, a named list of predictor variables (as rasters)}

\item{model_path}{character, the path for the models identified by
first doy of year for the 8-day week.}

\item{xcast_path}{character, the output path for the raster xcasts}
}
\value{
raster stack of predictions
}
\description{
Generate xcast rasters for a given species and year
}
