% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{plot_obs}
\alias{plot_obs}
\title{Plot the observations}
\usage{
plot_obs(x = read_obs(), raster_layer = NULL, ...)
}
\arguments{
\item{x}{table of observations}

\item{raster_layer}{a RasterLayer or NULL}

\item{...}{extra arguments for \code{\link[leaflet]{addRasterImage}}}
}
\value{
leaflet map object suitable for display
}
\description{
Plot the observations
}
